refactoring: (chunk) host-app
* 컴포넌트가 디렉토리 상에서 계층 구조를 가지도록 변경
* 컴포넌트 역할 및 파일 분리
