+++ frontend/host-app/src/components/CreateEventModal/CreateEventModalBody/HashTagArea/HashTagsField.js
@@ -4,7 +4,7 @@ import Chip from "@material-ui/core/Chip";
 import Paper from "@material-ui/core/Paper";
 import LocalOfferIcon from "@material-ui/icons/LocalOffer";
 import {Scrollbars} from "react-custom-scrollbars";
-import {SET_PROPERTY} from "./createEventModalActions.js";
+import {SET_PROPERTY} from "../../../../hooks/useCreateEventModal/createEventModalActions.js";
 
 const MyPaper = styled(Paper)({
 	marginTop: 10,
