+++ frontend/host-app/src/components/CreateEventModal/CreateEventModalBody/InputEventName.js
