+++ frontend/host-app/src/components/Event/EventCard/EventCardContent/EventCardContent.js
