+++ frontend/host-app/src/components/Event/EventCard/EventCardContent/EventCode.js
