+++ frontend/host-app/src/components/Event/EventCard/EventCardContent/EventName.js
