+++ frontend/host-app/src/components/Event/EventCard/EventIcon.js
