+++ frontend/host-app/src/components/EventSettingModal/EventSettingTab/ButtonField.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import useModal from "../../../customhook/useModal";
+import useModal from "../../../hooks/useModal";
 import ConfirmModal from "./ConfirmModal";
 
 const ButtonFieldStyle = styled.div`
