+++ frontend/host-app/src/components/Poll/PollContainer.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import {Button} from "@material-ui/core";
 import PollCard from "./PollCard";
 import NewPollModal from "./NewPollModal";
-import useModal from "../../customhook/useModal";
+import useModal from "../../hooks/useModal";
 import {useSocket, socketClient} from "../../libs/socket.io-Client-wrapper";
 
 // todo refactoring
