+++ frontend/host-app/src/graphql/gql.js
