+++ frontend/host-app/src/hooks/useSnackBar.js
